# university/src/main/resources/openapi.v1.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Timetable",
    "version": "1.0",
    "contact": {
      "name": "Taras Shablii"
    },
    "description": "Timetable App API spec"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1/timetable"
    }
  ],
  "paths": {
...
    "/auditoriums": {
      "get": {
        "summary": "List all auditoriums",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Auditorium"
                  }
                }
              }
            }
          }
        },
        "operationId": "get-auditoriums",
        "description": "get all auditoriums"
      },
      "post": {
        "summary": "Create new auditorium",
        "operationId": "post-auditoriums",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auditorium"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "description": "create new auditorium",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Auditorium"
              }
            }
          }
        }
      }
    },
    "/auditoriums/{id}": {
      "parameters": [
        {
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Retrieve auditorium",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auditorium"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-auditoriums-id",
        "description": "get single auditorium by id"
      },
      "put": {
        "summary": "Edit existing auditorium",
        "operationId": "put-auditoriums-id",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auditorium"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "description": "edit auditorium",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Auditorium"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove auditorium",
        "operationId": "delete-auditoriums-id",
        "responses": {
          "501": {
            "description": "Not Implemented"
          }
        },
        "description": "delete auditorium"
      }
    },
...
    "/schedules/{id}/available/auditoriums": {
      "parameters": [
        {
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Find available auditoriums",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Auditorium"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-schedules-id-available-auditoriums",
        "description": "get all available auditoriums for this schedule"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Auditorium": {
        "title": "Auditorium",
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name"
        ]
      },
...
    },
    "securitySchemes": {}
  }
}